refactor : 통합 테스트 로직 폴더 변경
